Create Upload All Master Data

+++ pkg/model/upload.go
@@ -3,5 +3,5 @@ package model
 type Upload struct {
 	File         string `json:"file"`
 	Table        string `json:"table"`
-	RowsAffected int64  `json:"rows_affected"`
+	RowsAffected int    `json:"rows_affected"`
 }
